Fix 2 compile errors: invalid operands to << and unary *
FUN_06004F14: Fixed operator precedence in pointer+shift expressions
  - Changed `puVar1 + x << 3 + 4` to `puVar1 + (x << 3) + 4`
  - The original generated invalid << on pointer type

FUN_06007590: Fixed dereference of cast-to-int expression
  - Changed `*((int)(int)param_1 << 1)` to `(*param_1) << 1`
  - Ghidra generated invalid dereference of integer value

Both were Ghidra decompilation artifacts generating invalid C.

+++ src/FUN_06007590.c
@@ -15,7 +15,7 @@ void FUN_06007590(param_1, param_2)
 
   uVar3 = *(unsigned short *)(0x060684EC + (unsigned int)param_1[1] * 2);
 
-  *(short *)(0x060684EC + (unsigned int)*((int)(int)param_1 << 1)) = (short)*(int *)0x0606A4F4;
+  *(short *)(0x060684EC + ((unsigned int)(*param_1) << 1)) = (short)*(int *)0x0606A4F4;
 
   *(int *)(puVar2 + *(int *)((int)(int)puVar1 << 3)) = *(int *)(puVar2 + (unsigned int)(uVar3 << 3));
